refactor: 패키지 이름 layered architecture에 맞게 변경
