+++ src/test/java/roomescape/presentation/ControllerTest.java
@@ -1,10 +1,10 @@
-package roomescape.controller;
+package roomescape.presentation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import roomescape.service.ReservationService;
+import roomescape.business.ReservationService;
 
 @WebMvcTest
 public abstract class ControllerTest {
